0.19.7 new package for Races + refactoring
