--- wfrp/src/main/java/com/warhammer/wfrp/dto/appearance/RaceDto.java
@@ -1,5 +1,0 @@
-package com.warhammer.wfrp.dto.appearance;
-
-public record RaceDto(String race) {
-
-}
