+++ wfrp/src/main/java/com/warhammer/wfrp/races/RaceDto.java
@@ -1,0 +1,5 @@
+package com.warhammer.wfrp.races;
+
+public record RaceDto(String race) {
+
+}
