+++ wfrp/src/main/java/com/warhammer/wfrp/races/RaceRepository.java
@@ -1,6 +1,5 @@
-package com.warhammer.wfrp.repository.race;
+package com.warhammer.wfrp.races;
 
-import com.warhammer.wfrp.model.race.Race;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RaceRepository extends JpaRepository<Race, Integer> {
